moving numpy frontend functions to correct files
